FAQ.md Updated
My name Is Phillip and I am a professional content/ advert writer.

I believe the current FAQ content is too confusing , and has a lot of grammatical errors.

So I decided to use my writing skill and contribute to the platform by correcting the sentences, rewriting some parts of the content and arrange them so it becomes more understandable and easier  to read.

Please merge to the main repository. 

Thanks.